extract check into hook
